+++ ionoscloud/resource_k8s_node_pool.go
@@ -159,6 +159,16 @@ func resourcek8sNodePool() *schema.Resource {
 					Type: schema.TypeString,
 				},
 			},
+			"labels": {
+				Type:     schema.TypeMap,
+				Optional: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+			},
+			"annotations": {
+				Type:     schema.TypeMap,
+				Optional: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+			},
 		},
 		Timeouts: &resourceDefaultTimeouts,
 	}
@@ -282,6 +292,24 @@ func resourcek8sNodePoolCreate(ctx context.Context, d *schema.ResourceData, meta
 			requestPublicIps = append(requestPublicIps, fmt.Sprint(publicIps[i]))
 		}
 		k8sNodepool.Properties.PublicIps = &requestPublicIps
+	}
+
+	labelsProp, ok := d.GetOk("labels")
+	if ok {
+		labels := make(map[string]string)
+		for k, v := range labelsProp.(map[string]interface{}) {
+			labels[k] = v.(string)
+		}
+		k8sNodepool.Properties.Labels = &labels
+	}
+
+	annotationsProp, ok := d.GetOk("annotations")
+	if ok {
+		annotations := make(map[string]string)
+		for k, v := range annotationsProp.(map[string]interface{}) {
+			annotations[k] = v.(string)
+		}
+		k8sNodepool.Properties.Annotations = &annotations
 	}
 
 	createdNodepool, _, err := client.KubernetesApi.K8sNodepoolsPost(ctx, d.Get("k8s_cluster_id").(string)).KubernetesNodePool(k8sNodepool).Execute()
@@ -467,6 +495,30 @@ func resourcek8sNodePoolRead(ctx context.Context, d *schema.ResourceData, meta i
 			diags := diag.FromErr(fmt.Errorf("error while setting lans property for k8sNodepool %s: %s", d.Id(), err))
 			return diags
 		}
+	}
+
+	labels := make(map[string]interface{})
+	if k8sNodepool.Properties.Labels != nil && len(*k8sNodepool.Properties.Labels) > 0 {
+		for k, v := range *k8sNodepool.Properties.Labels {
+			labels[k] = v
+		}
+	}
+
+	if err := d.Set("labels", labels); err != nil {
+		diags := diag.FromErr(fmt.Errorf("error while setting the labels property for k8sNodepool %s: %s", d.Id(), err))
+		return diags
+	}
+
+	annotations := make(map[string]interface{})
+	if k8sNodepool.Properties.Annotations != nil && len(*k8sNodepool.Properties.Annotations) > 0 {
+		for k, v := range *k8sNodepool.Properties.Annotations {
+			annotations[k] = v
+		}
+	}
+
+	if err := d.Set("annotations", annotations); err != nil {
+		diags := diag.FromErr(fmt.Errorf("error while setting the annotations property for k8sNodepool %s: %s", d.Id(), err))
+		return diags
 	}
 
 	return nil
@@ -684,6 +736,30 @@ func resourcek8sNodePoolUpdate(ctx context.Context, d *schema.ResourceData, meta
 
 			request.Properties.PublicIps = &requestPublicIps
 		}
+	}
+
+	if d.HasChange("labels") {
+		oldLabels, newLabels := d.GetChange("labels")
+		log.Printf("[INFO] k8s pool labels changed from %+v to %+v", oldLabels, newLabels)
+		labels := make(map[string]string)
+		if newLabels != nil {
+			for k, v := range newLabels.(map[string]interface{}) {
+				labels[k] = v.(string)
+			}
+		}
+		request.Properties.Labels = &labels
+	}
+
+	if d.HasChange("annotations") {
+		oldAnnotations, newAnnotations := d.GetChange("annotations")
+		log.Printf("[INFO] k8s pool annotations changed from %+v to %+v", oldAnnotations, newAnnotations)
+		annotations := make(map[string]string)
+		if newAnnotations != nil {
+			for k, v := range newAnnotations.(map[string]interface{}) {
+				annotations[k] = v.(string)
+			}
+		}
+		request.Properties.Annotations = &annotations
 	}
 
 	b, jErr := json.Marshal(request)